Added subsplit support
For folks who would like to track subsplits for things like picking up the flag, there's a setting for that. 
It's off by default, so anyone who wants to just use the autosplitter normally should select `Reset to Default` on the Settings page.
Easiest way to enable subsplits is to use the `Check All` button, and then uncheck the autoreset option only.